% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fxns.R
\name{add_periods_cols}
\alias{add_periods_cols}
\title{Add a year column as well as the appropriate day/week/month column}
\usage{
add_periods_cols(ident_dat = NULL, cadence = NULL)
}
\description{
Add a year column as well as the appropriate day/week/month column
}
