% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fxns.R
\name{AvgClim}
\alias{AvgClim}
\title{Climate Data Processing.}
\usage{
AvgClim(ident_data = NULL, clim_cols = NULL, all_years_ident = NULL,
  cadence = NULL)
}
\arguments{
\item{ident_data}{A mydataframe from 'mydata' for a single identifier's time-series mydata.}

\item{clim_cols}{The names of climate columns in ident_data (should match with column names in all_years_ident).}

\item{all_years_ident}{The output of get_disease_data containing all available years of mydata.}

\item{cadence}{An integer that specifies incidence cadence.}
}
\value{
ident_data  The same mydataframe, but with added historic averages for future climate dates.
}
\description{
This function is generally called by \code{\link{get.mysql()}} to fill future climate dates with appropriate historic averages.
After padding incidence with NAs, check climate data for NAs and replace with historic average.  Historic average to be calculated from all years available.
}
\examples{
require(DICE)

}
