% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fxns.R
\name{BuildDateVecs}
\alias{BuildDateVecs}
\title{Build A Date Vector}
\usage{
BuildDateVecs(cadence = cadence, start.date = start.date,
  end.date = end.date)
}
\arguments{
\item{cadence}{Numeric (1 = days, 2 = weeks, 3 = months and 4 = years)}

\item{start.date}{Start Date - Type Date}

\item{end.date}{End Date   - Type Date}
}
\value{
mydata A list with two/three elements: date, year and day/week/month
}
\description{
Given a cadence, start and end date build a vector of dates
}
\examples{
mydata <- BuildDateVecs(cadence = 2, start.date = "01-01-2015", end.date = "12-31-2015")
}
