% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fxns.R
\name{CadenceNum2Name}
\alias{CadenceNum2Name}
\title{Number to String Cadence Conversion}
\usage{
CadenceNum2Name(cadences = NULL)
}
\arguments{
\item{cadences}{An integer vector of cadences.}
}
\value{
cadence_names  A a vector of descriptive strings.
}
\description{
Convert number-based cadence representations to descriptive strings.
}
\examples{
require(DICE)
cadence_names = CadenceNum2Name(cadences=0:4)
}
