% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_fxns.R
\name{calc.epi.null}
\alias{calc.epi.null}
\title{Calculate the Historic NULL Model}
\usage{
calc.epi.null(mydata = NULL, all_years_epi = NULL, mod_id = NULL,
  fit_id = NULL)
}
\arguments{
\item{mydata}{- A dataframe with all the information for this \pkg{DICE} run}

\item{fit_id}{the abbreviated name for fit level regions}

\item{all_years_data}{the epi data for all years}

\item{model_id}{the abbreviated name for model level region}
}
\value{
A list with the NULL model values for the model and fit level data
}
\description{
\code{calc.null} Creates an epidemic NULL model using average monthly/weekly values from the previous ten years
}
\examples{
calc.epi.null(mydata = mydata, all_years_epi = all_years_epi, mod_id = mod_id, fit_id = fit_id)
}
