% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_fxns.R
\name{calc.mech.err}
\alias{calc.mech.err}
\title{Calculate Error in Mechanistic Fit/Forecast}
\usage{
calc.mech.err(tables = NULL, profiles = NULL, nfit = NULL,
  state_id = NULL)
}
\arguments{
\item{tables}{A list with all the data tables - both observed and fitted/forecasted}

\item{profiles}{An array with 10,000 selected dengue profiles from the MCMC fitting procedure}

\item{nfit}{An integer - the number of data points used in the fit}

\item{state_id}{The abbreviation of the state/region}
}
\value{
tables An updated tables list with the \code{nfit} row of the SEIR errors filled
}
\description{
\code{calc.mech.err} calculates the error:
mean absoulte error, mean relative error and mean error relative to the NULL model for each
month/week for which a fit/forecast was calculated
}
\examples{
calc.mech.err(tables=tables, profiles=profiles, nfit= nfit, state_id = state_id)
}
