% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_fxns.R
\name{calc.null.err}
\alias{calc.null.err}
\title{Calculate Error of Historic Null Model}
\usage{
calc.null.err(tables = NULL, nfit = NULL, state_id = NULL)
}
\arguments{
\item{tables}{A list with all the data tables - both observed and fitted/forecasted}

\item{nfit}{An integer - the number of data points used in the fit}

\item{state_id}{The abbreviation of the state/region}
}
\value{
tables An updated tables list with the \code{nfit} row of the NULL errors
- absolute and relative error
filled
}
\description{
\code{calc.null.err} calculates the error of the historic NULL model:
the weekly/monthly value is given by the ten year average value
}
\examples{
calc.null.err(tables=tables, nfit= nfit, state_id = state_id)
}
