% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plague_fxns.R
\name{calc.plague.err}
\alias{calc.plague.err}
\title{Calculate the lower/upper 95 percentile for the PLAGUE model
It puts a NA for the NULL model because we have data only for one season}
\usage{
calc.plague.err(profiles = NULL, mydata = NULL)
}
\description{
Calculate the lower/upper 95 percentile for the PLAGUE model
It puts a NA for the NULL model because we have data only for one season
}
