% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_dengue_fxns.R
\name{calc.sqldata.cor}
\alias{calc.sqldata.cor}
\title{Calculate Pearson Correlation Matrix Between Seasons}
\usage{
calc.sqldata.cor(mydata = NULL, all_years_epi = NULL, cases = NULL,
  nfit = NULL)
}
\arguments{
\item{mydata}{A dataframe with all the available data for this \pkg{DICE} run}

\item{all_years_epi}{the epi data for all years}

\item{cases}{AN array with the entire time series of incidence}

\item{nfit}{The number of data points used when calculating the correlation}

\item{epi}{the epi data for current season}
}
\value{
corrMAT An nyear x nyears matrix with the Pearson correlation values
}
\description{
\code{calc.sqldata.cor} Calculates an nyears by nyears correlation matrix for dengue incidence
}
\examples{
calc.cor(mydata= mydata, all_years_epi = al_years_epi, cases = all_years_epi$model$epi, nfit = nfit)
calc.cor(mydata= mydata, all_years_epi = al_years_epi, cases = all_years_epi$fit$epi[,iregion], nfit = nfit)
}
