% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_dengue_fxns.R
\name{calc.sqldata.dist}
\alias{calc.sqldata.dist}
\title{Calculate Euler Distance Matrix Between Seasons}
\usage{
calc.sqldata.dist(mydata = NULL, all_years_epi = NULL, cases = NULL,
  nfit = NULL)
}
\arguments{
\item{mydata}{A dataframe with all the available data for this \pkg{DICE} run}

\item{all_years_epi}{the epi data for all years}

\item{cases}{AN array with the entire time series of incidence}

\item{nfit}{The number of data points used when calculating the correlation}
}
\value{
distMAT An nyear x nyears matrix with Euler distance values
}
\description{
\code{calc.dist} Calculates an nyears X nyears Euler distance matrix for the incidence profiles
}
\examples{
calc.sqldata.dist(mydata= mydata, all_years_epi = al_years_epi, cases = all_years_epi$model$epi, nfit = nfit)
calc.sqldata.distr(mydata= mydata, all_years_epi = al_years_epi, cases = all_years_epi$fit$epi[,iregion], nfit = nfit)
}
