% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_fxns.R
\name{calc.stat.err}
\alias{calc.stat.err}
\title{Calculate Error in SARIMA Fit/Forecast}
\usage{
calc.stat.err(tables = NULL, nfit = NULL, state_id = NULL)
}
\arguments{
\item{tables}{A list with all the mydata tables - both observed and fitted/forecasted}

\item{nfit}{An integer - the number of data points used in the fit}

\item{state_id}{The abbreviation of the state/region}
}
\value{
tables An updated tables list with the \code{nfit} row of the SEIR errors filled
}
\description{
\code{calc.stat.err} calculates the error:
mean absolute error, mean relative error and mean error relative to the NULL model for each
month/week for which a fit/forecast was calculated
}
\examples{
calc.stat.err(tables=tables, nfit= nfit, state_id = state_id)
}
