% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{calcCPLRandomProfiles}
\alias{calcCPLRandomProfiles}
\title{Calculate Coupled Incidence Disease Profiles}
\usage{
calcCPLRandomProfiles(mydata = NULL, tab = NULL, tab.cpl = NULL,
  Rij = NULL, tps = NULL, nRnd = 100)
}
\arguments{
\item{mydata}{A list with the entire data available for this \code{DICE} run}

\item{tab}{The MCMC history of an indirect fit of the model using a coupled model
This  array includes all the parameters except the two that define the coupling matrix.}

\item{Rij}{A 2D array with the Euclidean distances between the centroids of regions \emph{i} and \emph{j}}

\item{tps}{A numeric array of days for the disease season}

\item{nRnd}{Integer  - the number of random profiles that will be calculated.  (Default is 100)}
}
\value{
profile a 3D  array with each row holding nRnd profiles for each of the fit regions.
}
\description{
Given an array with the MCMC history of a coupled \code{DICE} run
  randomly choose sets of parameters and use them to create disease incidence profiles for the fit regions.
  The code works on the level of the coupled fit regions.
}
\examples{
calcRandomProfiles{mydata = mydata, tab = tab, tab.cpl = tab.cpl, Rij = Rij,
tps = tps, nRnd = 100}
}
