% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fxns.R
\name{calcLLK}
\alias{calcLLK}
\title{Calculate the Likelihood of each Data Point}
\usage{
calcLLK(mod, dat)
}
\arguments{
\item{mod}{Numeric time series with the predicted number of cases}

\item{dat}{Integer time series with the observed number of cases}
}
\value{
myLLK  A vector with the likelihood value of each data point
}
\description{
Calculate the likelihood of each data point given two time-series: the observed one and predicted (dat and mod respectively).
}
\examples{
calcLLK(mod=model_rtn, dat = mydata$model$epi)
calcLLK(mod=model_rtn, dat = mydata$fit$epi[,iregion])
}
