% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{calcRandomProfiles}
\alias{calcRandomProfiles}
\title{Calculate Incidence Disease Profiles}
\usage{
calcRandomProfiles(sh = NULL, school = NULL, pop = NULL,
  tab = NULL, tps = NULL, nRnd = 100, cadence = 7, epi_model = 1)
}
\arguments{
\item{sh}{Numeric array with averaged specific humidity for the region}

\item{school}{Numeric array with the school vacation schedule for the region}

\item{pop}{Numeric - the region's total population}

\item{tab}{A numeric array with the MCMC history of all the parameters (both optimized and not)}

\item{tps}{A numeric array of days for the flu season - the day numbers correspond to the week numbers}

\item{nRnd}{Integer  - the number of random profiles that will be calculated.  (Default is 100)}

\item{cadence}{Integer - the cadence of the disease data}

\item{epi_model}{Integer 1 = SIR 2 = SEIR}
}
\value{
profile a 2D array with each row holding a single random disease incidence profile for the region
}
\description{
Given an array with the MCMC history of an uncoupled \code{DICE} run
  randomly choose sets of parameters and use them to create disease profiles.
  The code works on a level of single spatial region and supports both SIR and SEIR models.
}
\examples{
calcRandomProfiles{sh = mydata$model$sh, school = mydata$model$school,pop = mydata$model$pop, tab = tab,
tps = tps, nRnd = 100, cadence = 7}
}
