% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fxns.R
\name{cdc_download}
\alias{cdc_download}
\title{Pull ili mydata from CDC server (make 5 attempts)}
\usage{
cdc_download(CDCreg = NULL, years = NULL)
}
\description{
Pull ili mydata from CDC server (make 5 attempts)
}
