% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fxns.R
\name{CDCweek2date}
\alias{CDCweek2date}
\title{Convert a CDC week and year to Sunday's Date.}
\usage{
CDCweek2date(CDCweek, year)
}
\arguments{
\item{CDCweek}{An integer indicating the CDC week to be converted}

\item{year}{An integer specifying the year}
}
\value{
A `Date' class object containing the date of Sunday for the week and year provided.
}
\description{
CDC week numbering: first Wednesday of the year is in week 1.
}
\examples{
require(DICE)
CDCweek2date(10,2011)
}
