% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chop.R
\name{chop}
\alias{chop}
\alias{chop.rowwise_df}
\title{Chop a vector along the groups of a data frame}
\usage{
chop(data, x)

\method{chop}{rowwise_df}(data, x)
}
\arguments{
\item{data}{A tibble}

\item{x}{A vector, most often a column of `data`}
}
\value{
a list of slices of `x` with one element per group of `data`
}
\description{
Chop a vector along the groups of a data frame
}
\examples{
gdf <- iris \%>\%
  dplyr::group_by(Species)

chop(gdf, iris$Sepal.Length)
chop(gdf, 1:150)

}
