% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompPrimingPval.R
\name{CompPrimingPval}
\alias{CompPrimingPval}
\title{Assesses whether there is evidence for any priming in a cell population}
\usage{
CompPrimingPval(data.m, dice.o, nMC = 1000, ncores = 4)
}
\arguments{
\item{data.m}{The input data matrix, representing the normalized gene expression data matrix
over regulatory factors (rows) and single cells (columns).}

\item{dice}{The dice value as determined by CompDICE.}

\item{nMC}{Number of Monte-Carlo runs.}

\item{ncores}{Number of cores for parallelizing.}
}
\value{
A list of two elements

\item{pval}{Empirical P-value of significance.}

\item{null}{A vector with the null DICE values.}

}
\description{
This function evaluates whether the calculated DICE value of a cell population
is significantly lower than that of a randomized data matrix, indicating presence
of significant associations. This is mean to be applied to a multipotent cell population
to determine if there is significant priming.
}
\examples{


}
\references{
Qi Luo, Alok K Maity, Andrew E Teschendorff
\emph{Distance Covariance Entropy reveals primed states and bifurcation dynamics in single-cell RNA-Seq data}
Submitted.
}
