% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county_proto.R
\name{county_lnlike}
\alias{county_lnlike}
\title{Function to give a rough overall log likelihood for likleihood for 
To be efficient, needs to assume that all counties in the data are
in the model and that all season weeks in the data are in the model.
Assumes strictly that the ordering of the state county names is the 
same for the}
\usage{
county_lnlike(v_theta, mat_mod, df_dat, df_pop, seas = 2017,
  result = "num_spec", pseudo = FALSE)
}
\description{
Function to give a rough overall log likelihood for likleihood for 
To be efficient, needs to assume that all counties in the data are
in the model and that all season weeks in the data are in the model.
Assumes strictly that the ordering of the state county names is the 
same for the
}
