% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county_proto.R
\name{county_proto}
\alias{county_proto}
\title{A function to simulate a flu-like epidemic and be able to output 
which sub population infected which other sub-population}
\usage{
county_proto(df_pop, R0 = 1.8, Tg = 2.6, p_c = 1,
  trickle = 1/3e+08, kernPower = 0, d0 = 41, ws = 40, wf = 3,
  seedName = "First", seedSize = 50, dt = 1, nreals = 1,
  verbose = FALSE, beta_fudge = 1, week1 = 26)
}
\description{
A function to simulate a flu-like epidemic and be able to output 
which sub population infected which other sub-population
}
