% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fxns.R
\name{create_patch}
\alias{create_patch}
\title{Create a New Region}
\usage{
create_patch(mydata = NULL, RegState = NULL, fit_level = 3,
  mod_level = 2)
}
\arguments{
\item{mydata}{A mydata structure resulting from get.cdc.data() that contains all the regions listed in `RegState' in the $fit sub structure.}

\item{RegState}{A vector of strings specifying either the name or abbreviation of US regions or states.}

\item{fit_level}{An integer specifying the spatial-level of the elements of RegState. All elements of RegState must be from the same level.}

\item{mod_level}{An integer that defines the spatial-level of the created region.}
}
\value{
An updated version of `mydata' with only the RegState elements in $fit and the aggregated super-region in $model
}
\description{
Using the populations specified in `RegState` create a new region. If the user lists multiple entries to `RegState', this function will aggregate those regions into a new super-region.
}
\examples{
require(DICE)
mydata = get.DICE.data()
mydata2 = create_patch(mydata=mydata,RegState=c('R1','R2','R3'),fit_level=mydata$fit$level,mod_level=2)
}
