% Generated by roxygen2: do not edit by hand
\name{CreateDPTbin}
\alias{CreateDPTbin}
\title{Bins pseudotime into bins of equal cell numbers if possible}
\usage{
CreateDPTbin(dpt.v,minBins=7)
}
\arguments{
\item{dpt.v}{The input vector of pseudotime values.}

\item{minBins}{The number of bins required. The actual number will be at least as big as this one.}

}
\value{
A list with two elements

\item{binList}{A list of cellular indices, one for each bin, in increasing order of pseudotime.}

\item{idx}{A cellular index vector specifying the bin each cell falls into.}


}
\description{
Given as input a vector of pseudotime values, and the desired number of
bins, it places cells into bins of increased pseudotime values.
}
\examples{


}
\references{
Qi Luo, Alok K Maity, Andrew E Teschendorff
\emph{Distance Covariance Entropy reveals primed states and bifurcation dynamics in single-cell RNA-Seq data}
Submitted.
}
