% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fxns.R
\name{Date2CDCweek}
\alias{Date2CDCweek}
\title{Convert a Date to CDC Week and Year}
\usage{
Date2CDCweek(date = NULL)
}
\arguments{
\item{date}{A `Date' class object}
}
\value{
A list containing the CDC week and year that `date' falls in.
}
\description{
CDC week numbering: first Wednesday of the year is in week 1.
}
\examples{
require(DICE)
Date2CDCweek(as.Date('2011-03-15'))

}
