% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{dice}
\alias{dice}
\title{Lazily chop columns of a data frame}
\usage{
dice(data, callback = NULL)
}
\arguments{
\item{data}{A data frame}

\item{callback}{Callback function}
}
\description{
Lazily chop columns of a data frame
}
\examples{
e <- iris \%>\%
  dplyr::group_by(Species) \%>\%
  dice()

e$Sepal.Length
e$Sepal.Width

# Using callback ----------------------------------------------------------

e <- iris \%>\%
  dplyr::group_by(Species) \%>\%
  dice(callback = print)

chunks <- e$Sepal.Width

}
