\docType{data}
\name{diceBoot}
\alias{diceBoot}
\title{Bootstrapped DICE values for yangLiver dataset}
\format{A data matrix of DICE values defined over 1000 bootstraps and 7 developmental
  timepoints.
}

\usage{
data(diceBoot)
}
\description{
A data matrix of DICE values obtained by bootstrapping cells from each
of the 7 developmental timepoints. A total of 1000 bootstraps.
}

\details{
Upon loading the data object, we find the following:
\itemize{
  \item \code{diceBoot.m}: the bootstrapped DICE-valued matrix.
}
}

\references{

}
\keyword{datasets}
