% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fxns.R
\name{distance.matrix}
\alias{distance.matrix}
\title{Calculate a Euclidean Distance Matrix}
\usage{
distance.matrix(mydata = NULL)
}
\arguments{
\item{mydata}{The complete list with all the information for the run.}
}
\value{
A 2D matrix with the Euclidian distance between the centroids of the fit
  regions. (Units are km.)
}
\description{
Given a data frame with the centroid location (in lat/lon) of all the regions
  construct the corresponding Euclidian distance matrix.
}
\examples{
distance.matrix{mydata=mydata}
}
