% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindPrimedCells.R
\name{FindPrimedCells}
\alias{FindPrimedCells}
\title{Identifies the putative cells primed to a given lineage}
\usage{
FindPrimedCells(data.m, prf.lv, thE = 0)
}
\arguments{
\item{data.m}{The input data matrix, representing the normalized gene expression data matrix
over regulatory factors (rows) and single cells representing MPPs (columns).}

\item{prf.lv}{A list of vectors specifying the putative regulatory factors involved in priming of each lineage. Each entry of the list corresponds to one lineage and should have a name.}

\item{thE}{A scalar threshold on expression for calling a RF expressed in a given cell. By default 0.}
}

\value{
\item{primedCells.li}{A list of vectors specifying the indices for the putative primed cells of each lineage.}
}

\description{
If priming is present (as evaluated using \code{CompPrimingPval} function), and
having identified the putative priming factors, this function helps identify the
corresponding primed cells.
}
\examples{



}
\references{
Qi Luo, Alok K Maity, Andrew E Teschendorff
\emph{Distance Covariance Entropy reveals primed states and bifurcation dynamics in single-cell RNA-Seq data}
Submitted.
}
