% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindPrimingFactors.R
\name{FindPrimingFactors}
\alias{FindPrimingFactors}
\title{Identifies putative priming regulatory factor pairs}
\usage{
FindPrimingFactors(data.m)
}
\arguments{
  
\item{data.m}{The input data matrix, representing the normalized gene expression data matrix
over candidate regulatory factors (rows) and single cells (columns).}

}

\value{
A list with two elements:

\item{rankedLOO}{A vector of regulatory factors ranked in decreasing order of priming potential.}

\item{rankedPairs}{A table ranking the most likely pairs of regulatory
  factors involved in priming, with 1st column labeling the Spearman
  Rank Correlation and 2nd column labeling the corresponding P-value.}

}

\description{
If priming is present, as evaluated using `CompPrimingPval` function, then
one wishes to find the regulatory factors driving such priming. This function ranks
regulatory factors and regulatory factor pairs according to their likelihood of being
involved in priming.
}
\examples{

}

\references{
Qi Luo, Alok K Maity, Andrew E Teschendorff
\emph{Distance Covariance Entropy reveals primed states and bifurcation dynamics in single-cell RNA-Seq data}
Submitted.
}
