% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_user_data_fxns.R
\name{fit.user.data}
\alias{fit.user.data}
\title{Fit User Provided Incidence}
\usage{
fit.user.data(mydata = NULL, par_names = NULL, opt.list = NULL,
  run.list = NULL, iseed = NULL)
}
\arguments{
\item{mydata}{The \pkg{DICE} data list}

\item{par_names}{Array with the \pkg{DICE} parameter names}

\item{opt.list}{Array with TRUE/FALSE for each f the parameters}

\item{run.list}{A list with MCMC parameters}

\item{iseed}{Integer - seed for RNG. Default is NULL}
}
\value{
results A list with all the MCMC fitting/forecasting results and the input mydata list
}
\description{
\code{fit.user.data} fits and forecasts user provided incidence using an MCMC procedure
and a compartmental S-I-R orr S-E-I-R model
}
\examples{
results <- fit.user.data(mydata = mydata, par_names = par_names, 
opt.list = opt.list, run.list = run.list, iseed = 12345)	
}
