% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who_fxns.R
\name{fitOnePatchWHO}
\alias{fitOnePatchWHO}
\title{Driver Routine for a single region - WHO Data}
\usage{
fitOnePatchWHO(mydata = NULL, all_years_epi = all_years_epi,
  opt.list = NULL, run.list = NULL, ireal = 1, iseed = NULL)
}
\arguments{
\item{mydata}{A complex list with all available mydata for a given flu season model}

\item{all_years_epi}{- the entire history of incidence for this region}

\item{opt.list}{A Logical list with TRUE or FALSE values for all the parameters supported by \code{DICE}.   These values are set based on the user chosen model
for the basic reproduction value.}

\item{run.list}{A list with parameters needed for the MCMC procedure}

\item{ireal}{Integer - the MCMC chain number.  Default is 1.
#' @param iseed An integer used to set the random-number-generator seed. When not specified, the seed is generated randomly. Setting the seed to a known integer allows an MCMC chain to be reproducible.}
}
\value{
A list with the following arguments:
\describe{
\item{model_rtn}{The best result of the MCMC procedure for the model data}
\item{model_profile}{Randomly selected results from the MCMC procedure of directly fitting the model data}
\item{tab.model}{The MCMC history of the direct fit to the model data}
}
}
\description{
An MCMC fit of the model WHO data.
}
\examples{
fitOnePatchWHO{mydata = mydata, opt.list = opt.list, run.list = run.list, ireal = ireal, iseed = iseed}
}
