% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarima_fxns.R
\name{fitSARIMA}
\alias{fitSARIMA}
\title{SARIMA Fitting to Disease Data}
\usage{
fitSARIMA(mydata = NULL, all_years_epi = NULL, arima_model = NULL,
  auto_arima_model = NULL, device = "png")
}
\arguments{
\item{mydata}{- dataframe with all the data for this \pkg{DICE} run}

\item{all_years_epi}{the epi data for all years}

\item{arima_model}{List of ARIMA model parameters: list(p=, d=, q=, P=, D, Q=) can be set to NULL to trigger the}

\item{auto_arima_model}{A list of upper limit values for the ARIMA parameters.  This list is created by the
\code{DICE} code ONLY if arima_model is not set by the user}

\item{device}{- 'pdf' or 'png'. Default is 'png'
\code{auto.arima} process}
}
\value{
A list of tables with the results for the model (direct and aggregate) and fit regions
}
\description{
\code{fitSARIMA} fits the model and fit level regions with a SARIMA model (either chosen by the user or using \pkg{auto.arima})
and calls the plotting function
The model level is fitted both directly and as the weighted average of the fit level regions
}
\examples{
fitSARIMA((mydata = mydata, all_years_epi = all_years_epi, arima_model = arima_model,
auto_arima_model = NULL, covar = covar, covar_lag = covar_lag, run.list = run.list)
}
