% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sars_fxns.R
\name{fitSARS}
\alias{fitSARS}
\title{Driver Routine - Fitting a Single Plague Incidence Profile}
\usage{
fitSARS(mydata = NULL, all_years_epi = NULL, opt.list = NULL,
  run.list = NULL, ireal = 1, iseed = NULL)
}
\arguments{
\item{mydata}{A list containing all available information for the region: incidence, specific-humidity, school schedule, population etc.}

\item{opt.list}{A Logical list with TRUE or FALSE values for all the parameters supported by \code{DICE}.
These values are set based on the user chosen model for the compartmental model and the force of infection.}

\item{run.list}{A list with parameters needed for the MCMC procedure}

\item{ireal}{Integer - the MCMC chain number.  Default is 1.}

\item{iseed}{An integer used to set the random-number-generator seed. When not specified, the seed is generated randomly.
Setting the seed to a known integer allows an MCMC chain to be reproducible.}
}
\value{
A list with the following arguments:
\describe{
\item{model_rtn}{The best result of the MCMC procedure for the data}
\item{model_profile}{Randomly selected results from the MCMC procedure of fitting the data}
\item{tab.model}{The MCMC history of fitting the data}
}
}
\description{
A driver for fitting a single cholera incidence profile of a region/patch. The R code calls a Fortran routine which
uses an MCMC procedure to maximize the likelihood of the solution using and seir mode.
}
\examples{
fitPlague{mydata = mydata, opt.list = opt.list, run.list = run.list, ireal = ireal, iseed = iseed}
}
