% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{fitSingle}
\alias{fitSingle}
\title{Driver Routine - Uncoupled Spatial Model}
\usage{
fitSingle(mydata = NULL, all_years_epi = NULL, opt.list = NULL,
  run.list = NULL, ireal = 1, iseed = NULL)
}
\arguments{
\item{mydata}{A complex list with all available data for a given season model/fit spatial levels and data type}

\item{all_years_epi}{A complex list with all available data for all seasons model/fit spatial levels and data type}

\item{opt.list}{A Logical list with TRUE or FALSE values for all the parameters supported by \code{DICE}.   These values are set based on the user chosen model
for the basic reproduction value.}

\item{run.list}{A list with parameters needed for the MCMC procedure}

\item{ireal}{Integer - the MCMC chain number.  Default is 1.
#' @param iseed An integer used to set the random-number-generator seed. When not specified, the seed is generated randomly. Setting the seed to a known integer allows an MCMC chain to be reproducible.}
}
\value{
A list with the following arguments:
\describe{
\item{model_rtn}{The best result of the MCMC procedure for the model data}
\item{model_profile}{Randomly selected results from the MCMC procedure of directly fitting the model data}
\item{tab.model}{The MCMC history of the direct fit to the model data}
\item{fit_rtn}{The best result for indirectly fitting the model data using the fit regions}
\item{fit_profile}{Randomly selected results from the MCMC procedure of indirectly fitting the model data using the fit regions}
\item{tab.fit}{The MCMC history of indirectly fitting the model data using the fit regions}
}
}
\description{
A spatially uncoupled MCMC fit of the model data. The code first fits the model data directly and then
  fits each of the sub-regions sequentially - minimizing the likelihood of each one. The final indirect
  model fit is obtained as a weighted sum of these individual fits with the weights given by the relative
  population of each region. The data can be either cdc or gft data, and the model/fit data should have
  different spatial scales. For example in the case of cdc/gft data: the model can be national and the fit are
  the ten HHS regions. Or the model can be an HHS region and the fit are the states in that region.
}
\examples{
fitSingle{mydata = mydata, opt.list = opt.list, run.list = run.list, ireal = ireal, iseed = iseed}
}
