% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flights}
\alias{flights}
\title{Parameters coming from Gaussian fits to posterior distributions of fits to the CDC's ILI data}
\format{An object of class \code{data.frame} with 166 rows and 167 columns.}
\usage{
flights
}
\description{
Commute Rate by Country Based on Flight Volume Data
}
\details{
A data frame with 166 rows and 167 columns (first column should be ignored) denoting the normalized flight volume from origin (row) to all other countries,
Diagonal elements are \emph{NOT} included.  The three-letter ISO3 name convention is used to describe the countries.
}
\section{Warning}{

\bold{This Data Set is Not Currently Supported}
}

\keyword{datasets}
