% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fxns.R
\name{gammaEpi}
\alias{gammaEpi}
\title{Calculate the \eqn{Gamma} Function of Disease Number of Cases}
\usage{
gammaEpi(epi = NULL)
}
\arguments{
\item{epi}{A nperiods x nregions matrix of integers  with monthly/weekly/daily number of disease cases for each region}
}
\value{
gamaepi A matrix  of nperiods x nregions with the \eqn{\Gamma} function of monthly/weekly/daily number of disease cases in each region
}
\description{
Given a matrix or vector of integers with monthly/weekly/daily number of disease cases calculate for each spatial region, the \eqn{\Gamma} function of this matrix/vector.
This is done to save computational time during the MCMC procedure: the \eqn{\Gamma} function of the number of cases is needed when evaluating the maximum likelihood.
}
\examples{
gammaEpi{epi = mydata$model$epi$}
gammaEpi{epi = mydata$fit$epi$  }
}
