% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fxns.R
\name{get.cdc}
\alias{get.cdc}
\title{Retrieve incidence data from CDC server.}
\usage{
get.cdc(all_years = NULL, mod_level = 2, fit_level = 3,
  mod_name = c(NAME_2 = "USA"), myDB = NULL)
}
\arguments{
\item{all_years}{Data structure resulting from a call to \code{\link{get_disease_data}}}

\item{mod_level}{An integer describing the spatial level of the model data.(Default value is 2)  Levels: 0-Global, 1-Continent, 2-Country, 3-Region, 4-State, 5-County, 6-City.  \pkg{dice} currently has data at levels 2-4 for CDC and GFT.}

\item{fit_level}{An integer describing the spatial level of the fits used to construct the model-level profile/forecast (Default value is 3, must be >= mod_level).}

\item{mod_name}{A named-vector of character strings that specify which region is to be \strong{model}ed.  In other words, \code{mod_name} specifies the country, region, state, etc. of the mod_level region.  \code{mod_name} should be of the form \code{mod_name = c(NAME_2='a', NAME_3='b',..., NAME_i='x'} where i=mod_level and 'a', 'b',...,'x' are the appropriate level names. NAME_i='x' also accepts abbreviations.  Choose appropriate names from \code{\link{diceData}}.  For example, mod_name=c(NAME_2='United.States',NAME_3='Region4',NAME_4='North.Carolina') and mod_level=4 specifies North Carolina. To achieve the same result, use all abbreviations mod_name=c(NAME_2='USA',NAME_3='R4',NAME_4='NC') or a mix of names and abbreviations mod_name=c(NAME_2='USA',NAME_3='Region4',NAME_4='NC')}

\item{myDB}{A connection to one of the databases.  Generally opened using \code{\link{OpenCon}}.}
}
\description{
This function is generally called by \code{\link{get.DICE.data}}.
Augment/overwrite all_years with incidence data retrieved from the CDC server.  Also augment/overwrite specific humidity with NASA data retrieved from the Predictive Science server.
}
