% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_fxns.R
\name{get.first.year}
\alias{get.first.year}
\title{Find the First Year for the Data}
\usage{
get.first.year(mydata = NULL)
}
\arguments{
\item{mydata}{- A dataframe with all the information for this \pkg{DICE} run}
}
\value{
StartYear - a list with first year of data for this data type: both model level and fit level
}
\description{
This function is called after mydata was obtained to determine the first year of data for this data type
}
\examples{
FirstYear = get.first.year(mydata)
}
