% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fxns.R
\name{get.FitLevelName}
\alias{get.FitLevelName}
\title{Spatial Name for Fit Level}
\usage{
get.FitLevelName(iFit = 3)
}
\arguments{
\item{iFit}{An integer (Currently support 2-4 for cdc and gft data)}
}
\value{
A character string with the corresponding spatial Name: earth, continent, country, hhs_region, state, county and city for iFit between 0 and 6
}
\description{
Given a spatial Fit level this functions returns the corresponding spatial name. It is called by \code{\link{get.DICE.data}}. It assigns the correct spatial name to the spatial level
}
\examples{
get.FitLevelName(iFit=3)

get.FitLevelName(iFit=4)
}
