% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_fxns.R
\name{get.flu_poc.da}
\alias{get.flu_poc.da}
\title{Augment an influenza point-of-care data Time Series}
\usage{
get.flu_poc.da(mydata = NULL, all_years_epi = NULL)
}
\arguments{
\item{mydata}{A dataframe with all the available for this \pkg{DICE} run}
}
\value{
- A mydata data frame with the additional augmented data and its weight
}
\description{
\code{get.flu_poc.da} is a wrapper for the data augmentation procedure
Based on the value of the parameter 'da' it calls the function that augments
with the historic null model (da=1) or the most similar season (da=2).
If da=0 there is no data augmentation
}
\examples{
mydata <- get.flu_poc.da(mydata = mydata, all_years_epi=all_years_epi)

}
