% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_fxns.R
\name{get.flu_poc.future.average.data}
\alias{get.flu_poc.future.average.data}
\title{Data Augmentation Using the Most Similar Season}
\usage{
get.flu_poc.future.average.data(mydata = NULL, all_years_epi = NULL)
}
\arguments{
\item{mydata}{A dataframe with all the available for this \pkg{DICE} run}

\item{all_years_epi}{A dataframe with the entire history of disease incidence}
}
\value{
An updated mydata structure that has the future data and has weights for each week of data
}
\description{
given the model Point-of-Care data, use the historic average to augment the data
This function is needed when running in a forecast mode using data augmentation
}
\examples{
prior = get.flu_poc.future.average.data(mydata = mydata, all_years_epi = all_years_epi)

}
