% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_fxns.R
\name{get.future.average.data}
\alias{get.future.average.data}
\title{Data Augmentation Using the Historic NULL Model}
\usage{
get.future.average.data(mydata = NULL)
}
\arguments{
\item{mydata}{- A dataframe with all the available data for this \pkg{DICE} run}
}
\value{
An updated mydata structure that has the future data and has weights for each week of data
}
\description{
Use the average weekly incidence value for each region to augment the data
This function is needed when running in a forecast mode using data augmentation
}
\examples{
prior = get.future.average.data(mydata = mydata)
}
