% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fxns.R
\name{get.ModelLevelName}
\alias{get.ModelLevelName}
\title{Spatial Name for Model Level}
\usage{
get.ModelLevelName(mod_level = 2, RegState = "usa", NAME_4 = NULL,
  NAME_5 = NULL)
}
\arguments{
\item{mod_level}{An integer 2-4 for cdc and gft data}

\item{RegState}{A string, USA, or in case of mod_level=2, or region number of mod_level=3}

\item{NAME_4}{Relevant for mod_level = 4 and up.  State name}

\item{NAME_5}{Relevant for mod_level = 5 - county name}
}
\description{
Given a spatial Model level and the corresponding RegState value this function returns the corresponding spatial name
}
