% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_fxns.R
\name{get.prior}
\alias{get.prior}
\title{Find the Most Similar Season}
\usage{
get.prior(mydata = NULL)
}
\arguments{
\item{mydata}{A dataframe with all the available for this \pkg{DICE} run}
}
\value{
A list with the most similar season year for each of the fit and model regions
}
\description{
Given the fit and model ILI data find the most similar season for each of the fit and model regions
}
\details{
This function is needed when running in a forecast mode using a prior
}
\examples{
prior = get.prior(mydata = mydata)
}
