% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_fxns.R
\name{get.sd.da}
\alias{get.sd.da}
\title{Augment a San Diego Time Series}
\usage{
get.sd.da(mydata = NULL)
}
\arguments{
\item{mydata}{A dataframe with all the available for this \pkg{DICE} run}
}
\value{
- A mydata data frame with the additional augmented data and its weight
}
\description{
\code{get.sd.da} is a wrapper for the data augmentation procedure
Based on the value of the parameter 'da' it calls the function that augments
with the historic null model (da=1) or the most similar season (da=2).
If da=0 there is no data augmentation
}
\examples{
mydata <- get.cdc.da(mydata = mydata)

}
