% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fxns.R
\name{get.sh.nasa}
\alias{get.sh.nasa}
\title{Retrieve specific humidity data from data-server}
\usage{
get.sh.nasa(region = "all", sub_region = 1, start_year = 2000,
  start_week = 1, end_year = 2016, end_week = 10)
}
\arguments{
\item{region}{String specifying USA ('national'), CDC Region ('hhs'), US state ('state'), or all mydata regions ('all').}

\item{sub_region}{For region='hhs': Integer(s) indicating which of the 10 CDC regions are to be retrieved.  For region='state': a vector of state names to be returned.}

\item{start_year}{An integer - start year of the flu season}

\item{start_week}{An integer - starting CDC week of the flu season (default is 27)}

\item{end_year}{An integer  - end year of the flu season}

\item{end_week}{An integer - ending CDC week of the flu season (default is 26)}
}
\value{
a data frame with columns for $date, $week, $year, and whichever regions were requested.
}
\description{
Data is retrieved for the United States, HHS Regions, and States.
Attempt to pull CDC data from the CDC server. If that fails, pull from diceData.
}
\examples{
require(DICE)
sh = get.sh.nasa(region='hhs',sub_region=c(1,9),start_year=2015,end_year=2016,start_week=27,end_week=26)
sh2 = get.sh.nasa(region='state',sub_region=c('District.of.Columbia','Kansas'),start_year=2015,end_year=2016,start_week=27,end_week=26)
}
