% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fxns.R
\name{get_disease_data}
\alias{get_disease_data}
\title{Retrieve Incidence and Climate Data}
\usage{
get_disease_data(mod_level = 2, fit_level = 3, mod_name = c(NAME_2 =
  "BR"), years = "all", disease = "dengue",
  sql_data_source = integer(), daily_clim = FALSE, cad_clim = TRUE,
  mod_sum = FALSE, db_opts = list(DICE_db = "predsci", CDC_server = F))
}
\arguments{
\item{mod_level}{An integer describing the spatial level of the model data.(Default value is 2)  Levels: 0-Global, 1-Continent, 2-Country, 3-Region, 4-State, 5-County, 6-City.  \pkg{dice} currently has mydata at levels 2-3 for CDC and 2-4 for GFT.}

\item{fit_level}{An integer describing the spatial level of the fits used to construct the model-level profile/forecast (Default value is 3, must be >= mod_level).}

\item{mod_name}{A named-vector of character strings that specify which region is to be \strong{model}ed.  In other words, \code{mod_name} specifies the country, region, state, etc. of the mod_level region.  \code{mod_name} should be of the form \code{mod_name = c(NAME_2='a', NAME_3='b',..., NAME_i='x'} where i=mod_level and 'a', 'b',...,'x' are the appropriate level names. NAME_i='x' also accepts abbreviations.  Choose appropriate names from \code{\link{diceData}}.  For example, mod_name=c(NAME_2='United.States',NAME_3='Region4',NAME_4='North.Carolina') and mod_level=4 specifies North Carolina. To achieve the same result, use all abbreviations mod_name=c(NAME_2='US',NAME_3='R4',NAME_4='NC') or a mix of names and abbreviations mod_name=c(NAME_2='US',NAME_3='Region4',NAME_4='NC'). Unlike get.cdc.data(), get_disease_data() expects ISO2 country abbreviations.}

\item{years}{An integer vector - all years to be included.  If a value of "all" is passed, the function returns all available dates.}

\item{disease}{A character string - for example 'flu' or 'dengue'}

\item{sql_data_source}{An integer - designates which mydata source is to be used.}

\item{daily_clim}{Logic value determining if daily climate data is retrieved.  (Note: climate data averaged to the incidence-cadence is always returned.)}

\item{mod_sum}{Logic value indicating if model data will be aggregated from fit data.  If so, get_disease_data will not attempt to return model-level data.}
}
\value{
out  A list with a lookup table 'attr' and then climate data organized by cadence, spatial patch, and then data_source.
}
\description{
This function retrieves incidence and climate data from the MySQL database. It is generally called by \code{\link{get.DICE.data}}.
}
\examples{
require(DICE)
out = get_disease_data(mod_level = 2, fit_level = 4, mod_name = c(NAME_2='BR'), years="all", disease="dengue", sql_data_source=1)
}
