\docType{data}
\name{humanTF}
\alias{humanTF}
\title{List of human regulatory factors}
\format{Gene Symbols of 1536 human regulatory factors derived from the
  Molecular Signatures Database (MSigDB)
}

\usage{
data(humanTF)
}
\description{
A vector containing the official gene symbols of 1536 human regulatory factors
}

\details{
Upon loading the data object, we find the following:
\itemize{
  \item \code{humanTF.v}: the vector of gene symbols of human regulatory factors
}
}
\references{

}
\keyword{datasets}
