% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county_proto.R
\name{load_quidel_proto_data}
\alias{load_quidel_proto_data}
\title{So had a look at the data that Jamie extracted
Construct the state county combo
Have a quick look at the data 
Load up the US county data and create the right table
Check for counties in the data that are not in the county data table
and then extract them from the quidel data so we have at least a consistent 
pair of datasets. Need to look at the bad names to see if they are obviously 
in Alaska or something like that
Reduce the quidel data to only those with county names that are in our list of counties
Merge on the population sizes from the county database and sort the data}
\usage{

  load_quidel_proto_data(dataDir = "~/Dropbox/shares/pete_LEPR03/data/quidel/",
  week1 = 26)
}
\description{
So had a look at the data that Jamie extracted
Construct the state county combo
Have a quick look at the data 
Load up the US county data and create the right table
Check for counties in the data that are not in the county data table
and then extract them from the quidel data so we have at least a consistent 
pair of datasets. Need to look at the bad names to see if they are obviously 
in Alaska or something like that
Reduce the quidel data to only those with county names that are in our list of counties
Merge on the population sizes from the county database and sort the data
}
