% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fxns.R
\name{makeDir}
\alias{makeDir}
\title{Create a sub-directory}
\usage{
makeDir(subDir = NULL)
}
\arguments{
\item{subDir}{String - the sub-directory name for all output files of the run. Default it 'output'}
}
\description{
Create a sub-directory for a \pkg{DICE} run under the current working directory.   If this sub-directory already exists it does NOT overwrite it.
}
\examples{
makeDir{subDir='output'}
}
