% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_output_fxns.R
\name{mcmc.multi.write}
\alias{mcmc.multi.write}
\title{Write the MCMC History of a Coupled \pkg{DICE} Run}
\usage{
mcmc.multi.write(tab.model = NULL, tab.fit = NULL, tab.cpl = NULL,
  opt.list = NULL, opt.cpl = NULL, run.list = NULL, mydata = NULL,
  imask = NULL, cpl_imask = NULL, ireal = 1)
}
\arguments{
\item{tab.model}{The MCMC history of the direct fit of the model data}

\item{tab.fit}{The MCMC history of an indirect fit of the model using a coupled model.
This  array includes all the parameters except the two that define the coupling matrix.}

\item{tab.cpl}{The MCMC history of the two parameters that help define the coupling matrix:
the saturation distance and the distance power.}

\item{opt.list}{A logical list of all the parameters \pkg{DICE} recognizes and
can optimize with TRUE/FALSE}

\item{opt.cpl}{A logical list for the two parameters that help define the coupling matrix.}

\item{run.list}{a list with parameters used for the MCMC procedure}

\item{mydata}{A dataframe with all the data available for this run}

\item{imask}{AN array of integers with +1/-1 values for parameters that are optimized (or not)}

\item{cpl_imask}{An array of length two with +1/-1 for the two parameters that help define
the coupling matrix}

\item{ireal}{Integer - the MCMC chain number}
}
\value{
err   Returns \eqn{err = 0}
}
\description{
Writes an RData file with the MCMC history for an coupled \pkg{DICE} run.
  The function also calculates and prints to the screen the statistics for all the
  parameters that were optimized and does a Gaussian fit to these parameters.
  The results of these fits are written to separate a csv file.
}
\examples{
mcmc.single.write{tab.model = tab.model, tab= tab , tab.cpl = tab.cpl,
opt.list = opt.list,run.list = run.list, mydata = mydata, imask = imask,
cpl_imask = cpl_imask, ireal = 1}
}
