% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_output_fxns.R
\name{mcmc.onepatch.write}
\alias{mcmc.onepatch.write}
\title{Write the MCMC History of a Run - Single Region}
\usage{
mcmc.onepatch.write(tab.model = NULL, opt.list = NULL,
  run.list = NULL, mydata = NULL, imask = NULL, ireal = 1)
}
\arguments{
\item{tab.model}{The MCMC history of the direct fit of the mydata}

\item{opt.list}{A logical list of all the parameters \pkg{DICE} recognizes and
can optimize with TRUE/FALSE}

\item{run.list}{a list with parameters used for the MCMC procedure}

\item{mydata}{A dataframe with all the data available for this run}

\item{imask}{An array of integers with +1/-1 values for parameters that are optimized (or not)}

\item{ireal}{Integer - the MCMC chain number}
}
\value{
err   Returns \eqn{err = 0}
}
\description{
Writes an RData file with the MCMC history for a \pkg{DICE} run on a single region/patch.
  The function also calcualtes and prints to the screen the statistics for all the
  parameters that were optimized and does a Gaussian fit to these parameters.
  The results of these fits are written to separate a csv file.
}
\examples{
mcmc.onepatch.write{tab.model = tab.model, opt.list = opt.list,
run.list = run.list, mydata = mydata, imask = imask, ireal = 1}
}
