% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_fxns.R
\name{multiplot}
\alias{multiplot}
\title{Setup a grid for multi-plots on one page using ggplot2}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{plotlist}{A list with a ggplot in each element}

\item{file}{filename for output}

\item{cols}{Integer - the number of columns for the grid
layout optional a 2D matrix with the plot number in each location
matrix has ncols and nrow}
}
\description{
Setup a grid for multi-plots on one page using ggplot2
}
