% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fxns.R
\name{mydata_New2Old}
\alias{mydata_New2Old}
\title{Data Format Conversion}
\usage{
mydata_New2Old(mydata_new = NULL)
}
\description{
Convert the dataframe format from the new one to the old one.
Can only be used by expert Users
}
