% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fxns.R
\name{OpenCon}
\alias{OpenCon}
\title{Open a Connection to MySQL Database.}
\usage{
OpenCon(sql_db = "PredSci")
}
\arguments{
\item{sql_db}{Either 'PredSci' or 'BSVE'}
}
\value{
myDB  A S4 object that inherits from DBIConnection.
}
\description{
This function uses guest credentials to open a read-only connection to the MySQL database.
}
\examples{
require(DICE)
myDB = OpenCon(sql_db='PredSci')
}
