% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fxns.R
\name{PadNAs}
\alias{PadNAs}
\title{Pad Time Series wit NA's}
\usage{
PadNAs(ident_data = NULL, cadence = NULL, end.date = NULL,
  nmod = NULL)
}
\arguments{
\item{ident_data}{A dataframe from 'mydata' for a single identifier's time-series mydata.}

\item{cadence}{An integer that specifies incidence cadence.}

\item{end.date}{A Date class variable specifying the last date to be forecast.}
}
\value{
ident_data  The same mydataframe, but with added cadence periods until 'end.date'.
}
\description{
The function \code{PadNAs} takes a single identifier time-series dataframe from a 'data' \pkg{DICE} structure and add dates through 'end.date'.
This function is generally called by get.mysql() to add forecast dates beyond the last climate data point.
}
\examples{
require(DICE)

}
