% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fxns.R
\name{plotDisease}
\alias{plotDisease}
\title{Plot The entire Time Series of the Disease Data}
\usage{
plotDisease(mydata = NULL, all_years_epi = NULL, run.list = NULL,
  device = NULL)
}
\arguments{
\item{mydata}{A dataframe with all the available data for this \pkg{DICE} run}

\item{all_years_epi}{the epi data for all years}

\item{run.list}{a list with various parameters for the run}

\item{device}{- 'pdf or 'png'.  Default is 'png'
plotDisease(mydata = mydata, all_years_epi = all_years_epi, run.list = run.list, device = 'pdf')}
}
\value{
err=0 if plots were created
}
\description{
\code{plotDisease} Plots the disease time series for the country/level data along with the historic
average
}
