% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fxns.R
\name{plotFitCDCPercentILI}
\alias{plotFitCDCPercentILI}
\title{Plot the Results of a \pkg{DICE} Run}
\usage{
plotFitCDCPercentILI(rtn = NULL, profile = NULL, model_rtn = NULL,
  model_profile = NULL, mydata = NULL, ireal = 1, run.list = NULL,
  idevice = 1)
}
\arguments{
\item{rtn}{An nweeks x nregion 2D  numeric array with the best fit for each region}

\item{profile}{A 3D numeric array holding random predictions for each of the fit regions based on the history of their MCMC chains.}

\item{model_rtn}{A 1D numeric array with the best direct prediction to the model region}

\item{model_profile}{A 2D numeric array with randomly chosen predicted profiles obtained by fitting the model region directly.}

\item{mydata}{A dataframe with all the data available for this \pkg{DICE} run}

\item{ireal}{Integer - the MCMC chain number}

\item{run.list}{A list with various run parameters}

\item{idevice}{Integer - the index of the device in the device array. Default is 1 - make only one format of plot results}
}
\value{
Returns \eqn{err = 0} if successful
}
\description{
Plot the results of an a coupled or uncoupled  \pkg{DICE} run. For each of the fit regions we plot the incidence of the region along with
our predictions for it based on randomly selected results from the  history of the MCMC chain of each region.  Using the predictions
for the fit regions we then show the results for the model region as a weighted sum of the fit regions.  The last panel
shows our prediction for the model region using a direct fit to the model data.  The function also writes a binary RData file with
all the profile predictions for the model and fit regions. Note that in the case of a coupled run the fit regions are never individually
optimized. It is their weighted sum that is optimized, with the weights given by the relative population of each fit region.
}
\examples{
plotFitCDCPercentILI{ rtn = rtn, profile = profile, model_rtn = model_rtn, model_profile = model_profile,
mydata = mydata, ireal = ireal, device = device, idevice = 1}
}
