% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fxns.R
\name{plotFitOnePatch.external}
\alias{plotFitOnePatch.external}
\title{Plot the Results of a \pkg{DICE} Run - Using an External User Provided Routine}
\usage{
plotFitOnePatch.external(model_rtn = NULL, model_profile = NULL,
  mydata = NULL, ireal = NULL, run.list = NULL, idevice = 1)
}
\arguments{
\item{model_rtn}{A 1D numeric array with the best direct prediction to the region}

\item{model_profile}{A 2D numeric array with randomly chosen predicted profiles obtained by fitting the mydata}

\item{mydata}{A dataframe with all the data available for this \pkg{DICE} run}

\item{ireal}{Integer - the MCMC chain number}

\item{run.list}{A list with various run parameters}

\item{idevice}{Integer - the index of the device in the device array. Default is 1 - make only one format of plot results}
}
\value{
Returns \eqn{err = 0} if successful
}
\description{
Plot the results of \pkg{DICE} run for a single region/patch. We show the observed %ILI (or number of cases) along with our fits and
if appropriate predictions. We show the best result and randomly selected results from the MCMC chain.
For now this is just a wrapper that calls the \pkg{DICE} \code{\link{plotFitOnePatch.ggplot2}} function
}
\examples{
plotFitOnePatch{model_rtn = model_rtn, model_profile = model_profile,
mydata = mydata, ireal = ireal, device = device, idevice = 1}
}
