% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_fxns.R
\name{plotHists.ggplot2}
\alias{plotHists.ggplot2}
\title{Plot Histograms of Predicted and Observed Peak Week and Value}
\usage{
plotHists.ggplot2(rtn = NULL, profile = NULL, model_rtn = NULL,
  model_profile = NULL, mydata = NULL, ireal = 1, run.list = NULL,
  idevice = 1)
}
\arguments{
\item{rtn}{A 1D numeric array with the best in-direct prediction to the model region}

\item{profile}{A 3D numeric array holding random predictions for each of the fit regions based on the history of their MCMC chains.}

\item{model_rtn}{A 1D numeric array with the best direct prediction to the model region}

\item{model_profile}{A 2D numeric array with randomly chosen predicted profiles obtained by fitting the model region directly.}

\item{mydata}{A dataframe with all the data available for this \pkg{DICE} run}

\item{ireal}{Integer - the MCMC chain number}

\item{run.list}{a list with various parameters for the run}

\item{idevice}{Integer - the index of the device in the device array. Default is 1 - make only one format of plot results}
}
\value{
Returns \eqn{err = 0} if successful
}
\description{
Plots two histogram files one for the observed and predicted peak week and the other for the
observed and predicted \% ILI value for all weeks included in the range of the number of weeks fitted
to the number of weeks of data.  The \% ILI is presented in bins of 0.5\%.The default is to
fit all the available data.  In each case we first plot the results for all the fit regions, followed by
the results for an indirect (coupled or uncoupled) fit to the model region, and the last panel is for
the direct fit to the model region.
}
\examples{
plotHists.ggplot2(rtn = rtn, profile = profile, model_rtn = model_rtn, model_profile = model_profile,
mydata = mydata, ireal = ireal, run.list = run.list, idevice = 1)
}
