% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fxns.R
\name{plotMCMC}
\alias{plotMCMC}
\title{Plot Posterior Distribution from MCMC Procedure}
\usage{
plotMCMC(mydata = NULL, tab.model = NULL, tab.fit = NULL,
  tab.cpl = NULL, opt.list = NULL, opt.cpl = NULL, run.list = NULL,
  imask = NULL, ireal = 1, idevice = 1)
}
\arguments{
\item{mydata}{A dataframe with all the available data for this \pkg{DICE} run}

\item{tab.model}{The MCMC history of the direct fit of the model data}

\item{tab.cpl}{The MCMC history of the two parameters that help define the coupling matrix:
the saturation distance and the distance power.}

\item{opt.list}{A logical list of all the parameters \pkg{DICE} recognizes and
can optimize with TRUE/FALSE}

\item{opt.cpl}{A logical list for the two parameters that help define the coupling matrix.}

\item{run.list}{a list with parameters used for the MCMC procedure}

\item{imask}{AN array of integers with +1/-1 values for parameters that are optimized (or not)}

\item{ireal}{- Integer, the MCMC chain number}

\item{tab}{The MCMC history of an indirect fit of the model using a coupled model.
This  array includes all the parameters except the two that define the coupling matrix.}
}
\value{
err=0 if plots were created
}
\description{
\code{plotMCMC} Creates a PDF file with plots of the posterior distributions of all the mechanistic model \
parameters the User has chosen to fit and of the AICc score
}
\examples{
plotMCMC(mydata=mydata, Tg = Tg, tab.model = tab.model, tab.fit= tab.fit, tab.cpl = tab.cpl,
opt.list=opt.list, opt.cpl = opt.cpl, run.list = run.list, imask = imask, ireal = ireal)
}
