% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fxns.R
\name{plotMECH}
\alias{plotMECH}
\title{Plot Mechanistic Results of Fitting/Forecasting the Disease Data}
\usage{
plotMECH(mydata = NULL, tables.mod = NULL, tables.fit = NULL,
  tables.agg.mod = NULL, mod_id = NULL, fit_id = NULL,
  ymax.input = NULL, ireal = NULL, run.list = NULL, idevice = 1)
}
\arguments{
\item{mydata}{- dataframe with all the data for this \pkg{DICE} run}

\item{tables.mod}{- a table with the data and the results for the model fit. It includes the mean and the 5-95\% percentile (as well as the error)}

\item{tables.fit}{- Optional a table with the data and the results for the fits at the fit_level.
It includes the mean and the 5-95\% percentile (as well as the error)}

\item{tables.agg.mod}{- Optional. A table with the aggregate of the uncoupled fit_level results present only in the case of an uncoupled run}

\item{mod_id}{The abbreviation of the states/regions-model level}

\item{fit_id}{The abbreviation of the states/regions-fit level}

\item{ymax.input}{Optional, Maximum value for y-axis in plots (numeric)}

\item{ireal}{- Numeric, the number of the MCMC chain}
}
\value{
err=0 if plots were created
}
\description{
\code{plotMECH} Creates a PDF file with the results of the MCMC fits to the model and
the fit level data
}
\examples{
plotMECH(mydata = mydata, tables.mod = tables.mod, tables.fit = NULL,
mod_id = mod_id, fit_id = fit_id, ymax.input = NULL, ireal = ireal)
}
