% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_user_data_fxns.R
\name{plotUserDataMECH}
\alias{plotUserDataMECH}
\title{Plot Mechanistic Results of Fitting/Forecasting User Incidence Data}
\usage{
plotUserDataMECH(mydata = NULL, tables.mod = NULL, mod_id = NULL,
  ireal = 1, run.list = NULL, idevice = 1)
}
\arguments{
\item{mydata}{- dataframe with all the data for this \pkg{DICE} run}

\item{tables.mod}{- a table with the data and the results for the model fit. 
It includes the mean and the 5-95\% percentile (as well as the error)}

\item{mod_id}{The abbreviation of the states/regions-model level}

\item{ireal}{- Numeric, the number of the MCMC chain, default is 1}

\item{idevice}{Integer the index in the device array, default is 1}
}
\value{
err=0 if plots were created
}
\description{
\code{plotUserDataMECH} Creates a PDF or PNG file with the results of the MCMC fits to 
incidence data
}
\examples{
plotUserDataMECH(mydata = mydata, tables.mod = tables.mod,
mod_id = mod_id, ireal = ireal, run.list = run.list, idevice = 1)
}
