% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_user_data_fxns.R
\name{plotUserDisease}
\alias{plotUserDisease}
\title{Plot The Time Series of the User Provided Incidence}
\usage{
plotUserDisease(mydata = NULL, run.list = NULL, device = "png")
}
\arguments{
\item{mydata}{A list with all the available data for this \pkg{DICE} run}

\item{run.list}{a list with various parameters for the run}

\item{device}{- 'pdf or 'png'.  Default is 'png'
plotUserDisease(mydata = mydata, run.list = run.list, device = 'pdf')}
}
\value{
err=0 if plot was created
}
\description{
\code{plotUserDisease} Plots the user provided disease time series
}
