% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{port1}
\alias{port1}
\title{#' Gaussian Fit Parameters for Posterior Distributions
#' 
#' Results of Gaussian fits (mean and standard deviation) for each of the ten HHS regions using each of the five models
#' supported by \code{dice}. We only fit the basic reproduction number \eqn{R_0} and the percent clinical \eqn{pC}.  
#' In both cases we do a Gaussian fit to the posterior distribution.
#' This is done for each season, for each region and for each model.
#' These are needed if the MCMC procedure is done using a prior.  
#' The mean and standard deviation are packed into a named list with two elements:}
\format{An object of class \code{SpatialPolygonsDataFrame} with 52 rows and 13 columns.}
\usage{
port1
}
\description{
#' \describe{
#'   \item{mean}{The mean value}
#'   \item{sigma}{The standard deviation}
#' }
#' In each case there are 550 rows and 5 columns and each entry in the fourth and fifth column defines the \emph{mean} or \emph{sigma} for \eqn{R0} and \eqn{pC} given a specific 
#' set of: 
#' \describe{
#'   \item{Region}{Starting year for the prior flu season}
#'   \item{year.start}{Starting year for the prior flu season}
#'   \item{model}{model number 1-5}
#' }
}
\details{
"postFit"
The GADM_2.8 level 1 spatial polygon shape file for the USA

The file was downloaded from GADM using the getData command:
port1 = suppressMessages(getData('GADM', country = 'USA', level = 1))
and then converted to 'rda' format so it can be loaded with the package when LazyData is set to yes
The port1 object has 13 entries for each polygon
\describe{
 \item{OBJECTID}{Integer, polygon number}
 \item{ID_0}{Integer, country ID number}
 \item{ISO}{String, Three letter ISO3 country code}
 \item{NAME_0}{String, Country Name}
 \item{ID_1}{Integer, level 1 of country-state for the USA}
 \item{NAME_1}{String, level 1 name, state for the USA}
 \item{HASC_1}{String, two letter country code, and two letter state code, eg. US.CA}
 \item{CCN_1}{ NA in the case of USA}
 \item{CCA_1}{'' in the case of USA}
 \item{TYPE_1}{what type is the level 1 - State or Federal District in the case of USA}
 \item{ENGTYPE_1}{what type is the level 1 - State or Federal District in the case of USA}
 \item{NL_NAME_1}{'' in the case of USA}
 \item{VARNAME_1}{String, another state name descriptor in the case of USA}
}
}
\keyword{datasets}
