% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fxns.R
\name{print.error.dice}
\alias{print.error.dice}
\title{Print an Error Message}
\usage{
\method{print}{error.dice}(err, ...)
}
\arguments{
\item{err}{A character string with the error message}
}
\value{
No return value. Quit \pkg{DICE} with the save option set to yes
}
\description{
Prints the error message and quits \pkg{DICE} with save option set to yes
}
\examples{
print.error.dice('For GFT mydata sets DICE only supports flu seasons 2003-2014')
}
