% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{priorTable}
\alias{priorTable}
\title{Pearson Correlation Table for \% ILI - New version}
\format{An object of class \code{matrix} with 660 rows and 12 columns.}
\usage{
priorTable
}
\description{
A numeric correlation table with 132 rows and 6 columns showing for each season and each HHS region the season that has the highest correlation with it.
This season is referred to as the 'prior season'
Correlation is also shown for the national data- which is denoted as region 11.
\describe{
  \item{Region}{Region name,  RegionX and national - United.States}
  \item{start}{Starting year for the flu season}
  \item{end}{Ending year for the flu season}
  \item{R0-mean}{Mean value for R0}
  \item{R0-SD}{Standard deviation for R0}
  \item{pC-mean}{Mean value for percent clinical, pC}
  \item{pC-SD}{Standard deviation for percent clinical, pC}
  \item{SH-mean}{Mean value for SH term}
  \item{SH-SD}{Standard deviation for SH term}
  \item{SV-mean}{Mean value for SV term}
  \item{SV-SD}{Standard deviation for SV term}
  \item{Cor}{The Pearson correlation between the two seasons}
  ...
}
}
\keyword{datasets}
