% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fxns.R
\name{psi.lnlike.deldeviance}
\alias{psi.lnlike.deldeviance}
\title{Calculate the Difference Deviance of Two models}
\usage{
psi.lnlike.deldeviance(modI, modJ, dat)
}
\arguments{
\item{modI}{Numeric time series with the predicted number of cases using model I}

\item{modJ}{Numeric time series with the predicted number of cases using model J}

\item{dat}{Integer time series with the observed number of cases}
}
\value{
deldevi the difference in deviance of the two models
}
\description{
Calculate the difference in deviance given three time-series: the observed one and two predicted ones (dat, modI and modJ respectively).
}
