% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fxns.R
\name{psi.lnlike.deviance}
\alias{psi.lnlike.deviance}
\title{Calculate the Deviance of a Model}
\usage{
psi.lnlike.deviance(mod, dat)
}
\arguments{
\item{mod}{Numeric time series with the predicted number of cases}

\item{dat}{Integer time series with the observed number of cases}
}
\value{
devi the deviance of the model
}
\description{
Calculate the deviance given two time-series: the observed one and predicted (dat and mod respectively).
}
\examples{
psi.lnlike.deviance(mod=model_rtn, dat = mydata$model$epi)
psi.lnlike.deviance(mod=model_rtn, dat = mydata$fit$epi[,iregion])

}
