% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_fxns.R
\name{psi.sim.modelA}
\alias{psi.sim.modelA}
\title{A Compartmental SIR Model}
\usage{
psi.sim.modelA(R0 = 1.4, Tg = 2.6, t0 = 20, seed = 1, N = 10000,
  tps = (0:52) * 7, reals = 20, dt = 0.1, stochastic = TRUE,
  plot = TRUE)
}
\arguments{
\item{R0}{Numeric - the basic reproduction number}

\item{Tg}{Numeric - the recovery rate in days}

\item{t0}{Numeric - the onset day of the epidemic}

\item{seed}{Integer - the initial number of cases}

\item{N}{Integer - the total population}

\item{tps}{a numeric array of time points in days}

\item{reals}{Integer - the number of requested realizations.}

\item{dt}{Numeric - the time-step for the event-driven procedure}

\item{stochastic}{Logical - if TRUE (default) use a stochastic event-driven procedure, otherwise use a deterministic procedure}

\item{plot}{Logical - if TRUE plot the results}
}
\value{
rtn   a numeric array with the weekly incidence of all requested realizations
}
\description{
An R code for generating SIR profile(s) using either a stochastic or deterministic event-driven procedure.
  The routine calculates the weekly number of cases and if requested plots them.
  When using a stochastic procedure the user should set the number of realization to be > 50.
  When using a deterministic procedure the user should only ask for only one realization, see examples below.
}
\examples{
psi.sim.modelA(R0 = 1.4, Tg = 2.6, t0 = 20, seed = 1, N = 1e4, tps = (0:52) * 7, reals = 100, dt = 0.2, stochastic = TRUE , plot = TRUE)
psi.sim.modelA(R0 = 1.4, Tg = 2.6, t0 = 20, seed = 1, N = 1e4, tps = (0:52) * 7, reals = 1  , dt = 0.2, stochastic = FALSE, plot = TRUE)
}
