% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fxns.R
\name{quanci}
\alias{quanci}
\title{Calculate the 0.75 confidence interval of an a distribution}
\usage{
quanci(x)
}
\arguments{
\item{x}{- Numeric, a one dimensional vector of numbers}
}
\value{
the 0.75 lower and upper bounds
}
\description{
Calculate the 0.75 confidence interval of an a distribution
}
