% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{group_by}
\alias{tibble}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{group_by}}, \code{\link[dplyr]{group_by}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{tibble}{\code{\link[tibble]{tibble}}}
}}

