% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll}
\alias{roll}
\title{Roll any number of dice}
\usage{
roll(times = 1, sides = 6, verbose = TRUE)
}
\arguments{
\item{times}{Number of times to roll}

\item{sides}{How many sides the dice should have}

\item{verbose}{Print information about the roll to the screen}
}
\description{
a function for rolling dice!
}
\examples{
roll()
roll(times = 2, sides = 20)

}
