% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_user_data_fxns.R
\name{runDICEUserData}
\alias{runDICEUserData}
\title{Main Driver for using \code{DICE} for Fitting User Provided Incidence}
\usage{
runDICEUserData(filename = "data.csv", pop = 10000, epi_model = 1,
  Tg = 3, sigma = NULL)
}
\arguments{
\item{pop}{- Integer population of the region for which incidence is provided}

\item{epi_model}{- integer 1 (SIR) or 2 (SEIR), default is}

\item{Tg}{- Numeric, generation time in days. Default is 3 days}

\item{sigma}{- inverse of of the latent period in days. Needed only for an SEIR model. Default NULL}

\item{filenmae}{String - input csv file name.  Default 'data.csv'}
}
\value{
results A list with the input and entire output of the run.
}
\description{
\code{runDICEUserData} reads in a user provided incidence file and uses that along with the user
provided population size and value for the generation time, Tg, and if relevant the latent period
sigma to fit, and if desired, forecast the incidence.
Data cadence is arbitrary but at most can be monthly
We support only S-I-R and S-E-I-R models for a single population with a fixed force of infection
The incidence file must have two columns: dates and cases
All other parameters are set by the code
}
\examples{
Run an SEIR model using the incidence file and assuming a population of 1 million people.
The generation time and latent period are set to 2.6 days and 3 days respectively

output <- runDICEUserData(filename = 'data.csv', pop = 1e6, epi_model = 2, Tg = 2.6, sigma = 3.)

Run an SIR model using the incidence file and assuming a population of 10,000 people.
The generation time is set to 3 days. (No need to define a latent period.)
output <- runDICEUserData(filename = 'data.csv', pop = 1e5, epi_model = 1, Tg = 3.)

}
