% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_fxns.R
\name{runSyntheticMulti}
\alias{runSyntheticMulti}
\title{Driver Routine Coupled Spatial Model - Synthetic Data}
\usage{
runSyntheticMulti(mydata = NULL, opt.list = NULL, run.list = NULL,
  ireal = 1, iseed = NULL)
}
\arguments{
\item{mydata}{A complex list with all the data for a single  season. Everything except the number of cases and \%ILI
is from the 'cdc' dataType and will not be replaced.  Only the number of cases and the \%ILI will be replaced
by syntethic profiles}

\item{opt.list}{A Logical list with TRUE or FALSE values for all the parameters supported by \pkg{DICE}.
The synthetic example has been set-up for model = 4}

\item{run.list}{A list with parameters needed for the MCMC procedure}

\item{ireal}{Integer - the MCMC chain number.  Default is 1.}

\item{iseed}{An integer used to set the random-number-generator seed. When not specified, the seed is generated randomly. Setting the seed to a known integer allows an MCMC chain to be reproducible.}
}
\value{
A list with the following arguments:
\describe{
\item{model_rtn}{The best result of the MCMC procedure for the model synthetic data}
\item{model_profile}{Randomly selected results from the MCMC procedure of directly fitting the model synthetic data}
\item{tab.model}{The MCMC history of the direct fit to the model synthetic data}
\item{fit_rtn}{The best result for fitting the model synthetic data using the coupled synthetic fit regions}
\item{fit_profile}{Randomly selected results from the MCMC procedure of fitting the model data using the coupled fit regions}
\item{tab.fit}{The MCMC history of fitting the model data using the coupled fit regions}
}
}
\description{
A spatailly coupled MCMC fit of synthetic data. The code first fits the model data directly and then
  fits it using a weighted coupled model of the regions. The data is synthetic (although it has the dataType 'cdc').
  This synthetic example has been set up for the case of fit_level = 3 and mod_level = 2, that is fit the national data using the ten
  coupled HHS regions
}
\examples{
runSyntheticMulti{mydata = mydata, opt.list = opt.list, run.list = run.list, ireal = 1}
}
