% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_fxns.R
\name{sample.from.prior}
\alias{sample.from.prior}
\title{Sample from a Prior Gaussian Distribution}
\usage{
sample.from.prior(mydata = NULL, model_par = NULL, fit_par = NULL,
  model_pmin = NULL, model_pmax = NULL, fit_pmin = NULL,
  fit_pmax = NULL, model_ymu = NULL, model_sigma = NULL,
  fit_ymu = NULL, fit_sigma = NULL)
}
\arguments{
\item{mydata}{A dataframe with all the information for this \pkg{DICE} run}

\item{model_par}{An array with initial guesses for the parameters for the model region}

\item{fit_par}{A matrix with initial guesses for the parameters for the fit regions}

\item{model_pmin}{An array with minimum values for all the model region parameters}

\item{model_pmax}{An array with maximum values for all the model region parameters}

\item{fit_pmin}{A matrix with minimum values for all the fit regions parameters}

\item{fit_pmax}{A matrix with maximum values for all the fit regions parameters}

\item{model_ymu}{An array with the prior mean values for the model region}

\item{model_sigma}{An array with the prior standard deviation values for the model region}

\item{fit_ymu}{A matrix with the prior mean values for the fit regions}

\item{fit_sigma}{An array with the prior standard deviation values for the model region}
}
\value{
An array and matrix of initial parameters for the MCMC procedure
}
\description{
\code{sample.from.prior} samples initial guesses for the MCMC procedure using
a prior Gaussian distribution for: pC, R0 and the parameters that control
the force of infection for the SH and/or SV models
}
\examples{
sample.from.prior(mydata = mydata, model_par = model_par, fit_par = fit_par,
model_pmin = model_pmin, model_pmax = model_pmax, fit_pmin = fit_pmin, fit_pmax = fit_pmax,
model_ymu = model_ymu, model_sigma = model_sigma, fit_ymu = fit_ymu, fit_sigma = fit_sigma)

model 1 (SH): pC, R0, SH
model 2 (SV): pC, R0, SV
model 3 (SH+SV): pC, R0, SH, SV
model 4: pC, R0
model 5: pC, R0
}
