% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarima_fxns.R
\name{saveARIMA}
\alias{saveARIMA}
\title{Save the Results of a SARIMA Fit}
\usage{
saveARIMA(all_years_epi = NULL, mydata = NULL, tables.mod = NULL,
  tables.fit = NULL, tables.agg.mod = NULL, arima_model = NULL,
  arima_model_all = NULL, covar = NULL, covar_lag = NULL)
}
\arguments{
\item{mydata}{- dataframe with all the data for this \pkg{DICE} run}

\item{tables.mod}{- a table with the data and the results for the direct Model fit
It includes the mean and the 5-95\% percentile (as well as the error)}

\item{tables.fit}{- Optional a table with the data and the results for the fits at the fit_level. It includes the mean and the 5-95\% percentile (as well as the error)}

\item{tables.agg.mod}{- A table with aggregate results of the fits}

\item{arima_model}{- A list with the user selection of the ARIMA model
If the user has not selected a model this will be NULL and arima_model_all
will have the models selected by auto.arima}

\item{arima_model_all}{- An array with details of the ARIMA model used for
model region, the fits regions and the aggregate (the last is only if the User
has chosen an ARIMA model for the run)}

\item{covar}{- String. Optional covariate variable for ARIMA fit.
'sh', 'precip' and 'temp' are currently supported. Default is NULL - no covariate}

\item{covar_lag}{- Numeric. Lag time for covariate variable in units of the cadence
of the data}
}
\value{
A list of tables with the results for the model (direct and aggregate) and fit regions
}
\description{
\code{saveARIMA} saves the results of a SARIMA fit to the data into an RData file
}
\examples{
saveARIMA(all_years_epi = all_years_epi, mydata = mydata, tables.mod = tables.mod, tables.fit = tables.fit,
tables.agg.mod = tables.agg.mod,arima_model = arima_model,
arima_model_all = NULL, covar = covar, covar_lag = covar_lag)
}
