% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fxns.R
\name{saveProfiles}
\alias{saveProfiles}
\title{Save the profiles for Model region
Save the profiles for Model region so that incidence plot can be re-created}
\usage{
saveProfiles(mydata = NULL, model_rtn = NULL, model_profile = NULL,
  rtn = NULL, profile = NULL, ireal = 1, run.list = NULL)
}
\arguments{
\item{mydata}{A dataframe with all the data available for this \pkg{DICE} run}

\item{model_rtn}{A 1D numeric array with the best direct prediction to the model region}

\item{model_profile}{A 2D numeric array with randomly chosen predicted profiles obtained by fitting the model region directly.}

\item{rtn}{An nweeks x nregion 2D  numeric array with the best fit for each region}

\item{profile}{A 3D numeric array holding random predictions for each of the fit regions based on the history of their MCMC chains.}

\item{ireal}{Integer - the MCMC chain number}

\item{run.list}{A list with various run parameters}

\item{profile}{A 3D numeric array holding random predictions for each of the fit regions based on the history of their MCMC 
chains.}
}
\value{
Returns \eqn{err = 0} if successful
}
\description{
Save the profiles for Model region
Save the profiles for Model region so that incidence plot can be re-created
}
\examples{
saveProfiles{ mydata = mydata, model_rtn = model_rtn, model_profile = model_profile, rtn = rtn, 
profile = profile,  ireal = ireal, run.list = run.list}	
}
