% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fxns.R
\name{saveProfilesOnePatch}
\alias{saveProfilesOnePatch}
\title{Save the profiles for Model region
Save the profiles for Model region so that incidence plot can be re-created}
\usage{
saveProfilesOnePatch(mydata = NULL, model_rtn = NULL,
  model_profile = NULL, model_profileB = NULL, ireal = 1,
  run.list = NULL)
}
\arguments{
\item{mydata}{A dataframe with all the data available for this \pkg{DICE} run}

\item{model_rtn}{A 1D numeric array with the best direct prediction to the model region}

\item{model_profile}{A 2D numeric array with randomly chosen predicted profiles obtained by fitting the model region directly.}

\item{model_profileB}{A 2D numeric array with randomly chosen predicted profiles for the Bacteria obtained by fitting the model region directly. Relevant only in the case of an SIRB model}

\item{ireal}{Integer - the MCMC chain number}

\item{run.list}{A list with various run parameters}
}
\value{
Returns \eqn{err = 0} if successful
}
\description{
Save the profiles for Model region
Save the profiles for Model region so that incidence plot can be re-created
}
\examples{
saveProfilesOnePatch{ mydata = mydata, model_rtn = model_rtn, model_profile = model_profile,
model_profileB = model_profileB, ireal = ireal, run.list = run.list}	
}
