% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fxns.R
\name{saveRData}
\alias{saveRData}
\title{Save a Binary RData file}
\usage{
saveRData(mydata = NULL, all_years_epi = NULL, run.input = NULL,
  ireal = 1)
}
\arguments{
\item{mydata}{- dataframe with all the data for this \pkg{DICE} run}

\item{all_years_epi}{- a dataframe with the entire incidence and climate history}

\item{run.input}{a list of all other input  needed for the run}

\item{ireal}{Integer indicating the MCMC chain number}
}
\value{
err   \eqn{err =0} if successful.
}
\description{
Saves a binary RData file with all the input parameters and the data for this run.
  Each MCMC chain saves its own RData file and the file name includes the MCMC chain number
  as the last digit before the .RData extension of the file name.
}
\examples{
saveRData(mydata = mydata, run.input= run.input,ireal = 1)
}
