% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fxns.R
\name{saveTables}
\alias{saveTables}
\title{Save Tables with Results of Fitting/Forecasting the Disease Data}
\usage{
saveTables(mydata = NULL, tables.mod = NULL, tables.fit = NULL,
  tables.agg.mod = NULL, mod_id = NULL, fit_id = NULL,
  run.list = NULL, ireal = 1)
}
\arguments{
\item{mydata}{- dataframe with all the data for this \pkg{DICE} run}

\item{tables.mod}{- a table with the data and the results for the model fit. It includes the mean and the 5-95\% percentile (as well as the error)}

\item{tables.fit}{- Optional a table with the data and the results for the fits at the fit_level.
It includes the mean and the 5-95\% percentile (as well as the error)}

\item{tables.agg.mod}{- Optional. A table with the aggregate of the uncoupled fit_level results present only in the case of an uncoupled run}

\item{mod_id}{The abbreviation of the states/regions-model level}

\item{fit_id}{The abbreviation of the states/regions-fit level}

\item{run.list}{A list with various run parameters}

\item{ireal}{- Numeric, the number of the MCMC chain}
}
\value{
err=0 if file was saved
}
\description{
\code{saveTables} Saves and RData binary files with tables of the forecast/fit results as well as the observations
the fit level data
}
\examples{
saveTables(mydata = mydata, tables.mod = tables.mod, tables.fit = NULL,
mod_id = mod_id, fit_id = fit_id, ireal = ireal)
}
