% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_mcmc_fxns.R
\name{set.cpl.imask}
\alias{set.cpl.imask}
\title{Set a Mask for the Coupling Parameters}
\usage{
set.cpl.imask(nparam = NULL, opt.list = NULL)
}
\arguments{
\item{nparam}{Integer - The number of parameters that help define the coupling matrix.   Currently two.}

\item{opt.list}{A named logical list with TRUE or FALSE values for the coupling parameters.}
}
\value{
imask An integer list of length nparam with +1 or -1 values
examples
set.cpl.imask{nparam=cpl_nparam,opt.list=opt.cpl}
}
\description{
Set +1 or -1 values for the two parameter that help define the
coupling matrix.    These are needed and called for only in the case
of a coupled \pkg{DICE} run.
}
