% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fxns.R
\name{set.iseed}
\alias{set.iseed}
\title{Set the seed for the RNG}
\usage{
set.iseed()
}
\arguments{
\item{none}{required}
}
\value{
An integer - seed for the RNG
}
\description{
Create the seed in the R code for the Random Number Generator
}
\examples{
set.iseed()
}
