% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_run_fxns.R
\name{set.opt.list}
\alias{set.opt.list}
\title{Create a Logical List  with all the \pkg{DICE} model parameters}
\usage{
set.opt.list(mydata = NULL, isingle = 1)
}
\arguments{
\item{isingle}{A Number:
0 - coupled run
1 - uncoupled run}

\item{Compartmental}{Model Number (1-5) \itemize{
\item1 - SIR
\item 2 - SEIR
\item 3 - (SIR)_H / (SI)_V
\item 4 - (SEIR)_H / (SEI)_V
 \item 5 - SIRB}}

\item{model}{Number - The model Number (1-5, default is 5) \itemize{
\item 1 - specific humidity (SH) only
\item 2 - school vacation (SV) only
\item 3 - both SH and SV
\item 4 - Fixed value of R(t)
\item 5 - A two-value model for R(t) }}
}
\value{
A logical list with all the \pkg{DICE} parameters properly set to TRUE or FALSE.
}
\description{
\pkg{DICE} has a list of model parameters it recognizes, for both uncoupled and coupled runs.
This function sets the values of these parameters to either TRUE or FALSE depending on the model number
Given a model number TRUE or FALSE values are assigned to all of these parameters.
In the case of a coupled run, the coupled logical list is also constructed.
}
\examples{
set.opt.list{mydata = mydata, isingle = 0}
set.opt.list{mydata = mydata, isingle = 1}

}
