% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_run_fxns.R
\name{set.param.cpl.list}
\alias{set.param.cpl.list}
\title{Create a List of Coupling Parameters that the Code Recognizes}
\usage{
set.param.cpl.list()
}
\value{
An array with parameter names - the order of parameters will set the order for the min/max arrays also
and the 'mask' for which parameters are optimized (or not)
}
\description{
Create a list with the coupling paramter names
}
\examples{
set.param.cpl.list()
}
