% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_run_fxns.R
\name{set.param.list}
\alias{set.param.list}
\title{Create a List of all Parameters the Code Recognizes}
\usage{
set.param.list(epi_model = 1)
}
\arguments{
\item{epi_model}{Integer Mechanistic Model Number: 1=SIR, 2=SEIR, 3=(SIR)_H/(SI)_V, 4=(SEIR)_H/(SEI)_V}
}
\value{
An array with parameter names - the order of parameters will set the order for the min/max arrays also
and the 'mask' for which parameters are optimized (or not)
}
\description{
\code{set.param.list} creates a list with all the parameters that the \pkg{DICE} code recognizes. the list is `epi_model' specific.  If the vector states are included it will have additional parameters.
}
\examples{
set.param.list(epi_model = 2)
}
