% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_run_fxns.R
\name{set.run.list}
\alias{set.run.list}
\title{Create a List of Parameters For MCMC Procedure}
\usage{
set.run.list(nreal = 1, nMCMC = 1e+05, nlines = NULL,
  device = "pdf", subDir = "output", plot = 1)
}
\arguments{
\item{nreal}{Integer  - The number of MCMC chains (default is 1)}

\item{nMCMC}{Integer - the number of steps in each MCMC chain (default is 1e5)}

\item{nlines}{Integer - the number of instances saved for the history of the chain (default is to save every 100 steps and not to exceed 1e4)}

\item{device}{String  - the device name for plotting.  Default is 'pdf' but we also support 'png'}

\item{subDir}{String  - the sub-directory name for all output files of the run. Default it 'output'}

\item{plot}{Character - TRUE, FALSE or EXTERNAL allows users to use their own plotting routines for the results (can also use 0, 1 and 2)}
}
\value{
A list packed with these parameters
}
\description{
Pack various parameters related to the MCMC procedure and plotting into a single list.
  These include: The number of MCMC chains, the number of MCMC steps in each chain, the number
  of instances saved for each chain and the device name for plotting the results.
}
\examples{
set.run.list{nreal = 1, nMCMC = 1e+05, nlines = 1e3, device = 'pdf'}
set.run.list{nreal = 3, nMCMC = 5e+06, nlines = 1e4, device = 'pdf'}
}
