% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sars_fxns.R
\name{set.sars.model}
\alias{set.sars.model}
\title{Setup of Parameters for an MCMC SEIR procedure}
\usage{
set.sars.model(mydata = NULL, par_names = NULL, est_bckgrnd = 1,
  opt.list = NULL, run.list = NULL)
}
\arguments{
\item{par_names}{A character array with seir parameter names}

\item{est_bckgrnd}{Numeric, estimated value of background cholera cases}

\item{opt.list}{A Logical list with TRUE or FALSE values for all the parameters
supported by \pkg{DICE}. These values are based on the model used for the basic
reproduction number}

\item{run.list}{A list with parameters needed for the MCMC procedure}
}
\value{
A list with min, max, step size and initial guess/default values for the sir parameters
}
\description{
\code{set.sars.model} Creates the arrays for SEIR modeling of the Plague data with min/max values, step size,
and initial guess/default values for all the parameters.
}
\examples{
set.plague.model(mydata = mydata, par_names = par_names, 
est_bckgrnd=1, run.list = run.list, opt.list = opt.list)
}
