% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_user_data_fxns.R
\name{set.user.data.model}
\alias{set.user.data.model}
\title{Setup of Parameters for an MCMC procedure}
\usage{
set.user.data.model(mydata = NULL, par_names = NULL, opt.list = NULL,
  run.list = NULL)
}
\arguments{
\item{par_names}{A character array with parameter names}

\item{opt.list}{A Logical list with TRUE or FALSE values for all the parameters
supported by \pkg{DICE}.}

\item{run.list}{A list with parameters needed for the MCMC procedure}
}
\value{
A list with min, max, step size and initial guess/default values for the sir parameters
}
\description{
\code{set.user.data.model} Creates the arrays for modeling of user 
provided incidence data with min/max values, step size,
and initial guess/default values for all the parameters.
}
\examples{
set.user.data.model(mydata = mydata, par_names = par_names,
run.list = run.list, opt.list = opt.list)
}
