% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_user_data_fxns.R
\name{set.user.data.opt.list}
\alias{set.user.data.opt.list}
\title{Create a Logical List  with TRUE/FALSE values for parameter optimization}
\usage{
set.user.data.opt.list(mydata = NULL)
}
\arguments{
\item{mydata}{- The \code{DICE}  data list}
}
\description{
\pkg{DICE} has a list of model parameters it recognizes, for both uncoupled and coupled runs.
This function sets the values of these parameters to either TRUE or FALSE for the simple
case of user provided data
}
\examples{
set.opt.list{mydata = mydata}

}
