% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_user_data_fxns.R
\name{set.user.data.param.list}
\alias{set.user.data.param.list}
\title{Short Parameter List - Fixed Force of Infection Case}
\usage{
set.user.data.param.list()
}
\value{
An array with parameter names - the order of parameters will set the order for the min/max arrays also
and the 'mask' for which parameters are optimized (or not)
}
\description{
\code{set.user.data.param.list} creates a list with parameters that the \pkg{DICE} code recognizes.
}
\examples{
set.user.data.param.list()
}
