% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_mcmc_fxns.R
\name{setup.coupling.mcmc}
\alias{setup.coupling.mcmc}
\title{Setup For Coupling Parameters}
\usage{
setup.coupling.mcmc(par_names_cpl = NULL, opt.list = NULL)
}
\arguments{
\item{par_names_cpl}{An array with the names of the coupling parameters}

\item{opt.list}{a logical list with TRUE or FALSE values for each}
}
\value{
coupling.list A named list with the following arguments:
\describe{
\item{parmin}{Minimum values for the two parameters}
\item{parmax}{Maximum values for the two parameters}
\item{pardx}{Step-size for the two parameters}
\item{par}{Initial values for the two parameters}
\item{logbase}{Base for log values - currently code assumes base 10}
\item{logvec}{Array of integers with 1, use log base, or 0 - do not}
\item{nparam}{Integer-the total number of parameters recognized by the \pkg{DICE} code}
\item{nopt}{Integer-the number of parameters that will be optimized}
}
}
\description{
For a coupled \pkg{DICE} run we setup all the required values for
the two parameters that help define the coupling matrix.  These
  include min/max and step size values, initial values, logscale
and number of parameters. The two parameters are: the saturation distance
  \eqn{s_d} in km, and the distance power \eqn{\gamma}.
}
\examples{
setup.coupling.mcmc{par_names_cpl = par_names_cpl, opt.list=opt.cpl}
}
