% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_mcmc_fxns.R
\name{setup.fit.mcmc}
\alias{setup.fit.mcmc}
\title{Setup parameters for MCMC Procedure - Fit Data}
\usage{
setup.fit.mcmc(mydata = NULL, run.list = NULL, opt.list = NULL,
  tps = NULL, par_names = par_names)
}
\arguments{
\item{mydata}{- dataframe with all the data for this \pkg{DICE} run}

\item{run.list}{A list with parameters needed for the MCMC procedure}

\item{opt.list}{A Logical list with TRUE or FALSE values for all the parameters
and find the proper values needed for this prior.
supported by \pkg{DICE}. These values are based on the model used for the basic
reproduction number.}

\item{tps}{A numeric array of days for the disease season - the day numbers are
consistent with the weeks/months}

\item{par_names}{- - An array with the all parameters ordered as required by DICE}
}
\value{
A list with the following  arguments:
\describe{
\item{par_min}{Minimum values for all the parameters supported by \pkg{DICE} for each region}
\item{par_max}{Maximum values for all the parameters supported by \pkg{DICE} for each region}
\item{pardx}{Step-size for MCMC for all the parameters supported by \pkg{DICE} for each region}
\item{par}{Initial values for all the parameters for each region}
\item{logbase}{Base for log values - currently code assumes base 10}
\item{nopt}{Integer-the number of parameters that will be optimized}
\item{tab}{A 2D numeric array with nlines and (nparam+1) columns used to store the MCMC history
  of all the parameters and the likelihood}
}
}
\description{
Setup all the required parameters for the MCMC procedure on the fit data.
  These include: The min/max and step size values for all the parameters,
  the initial values for all the parameters, a vector with the list of parameters
  that will be optimized, the total number of parameters and the number of
  parameters that will be optimized.  The code also allocates an array, tab,
  where the history of the MCMC chain is recorded.
}
\examples{
setup.fit.mcmc{mydata = mydata,
run.list = run.list, opt.list = opt.list, tps = tps, par_names = par_names}
}
