% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dengue_fxns.R
\name{subset.sql.mydata}
\alias{subset.sql.mydata}
\title{Subsets the Data}
\usage{
\method{subset}{sql.mydata}(years = NULL, year.start = 2010,
  cadence = "month", cadence.start = 1, ntps = 12, cases = NULL)
}
\arguments{
\item{years}{A vector of years}

\item{year.start}{Start year for subseting data}

\item{cadence}{String, the cadence of the data: month, week, day}

\item{cadence.start}{Start month/week for subseting data}

\item{ntps}{Length of subset data}

\item{cases}{A data frame with year, month/week, ndays and cases information}
}
\value{
cases A data frame of length ntps starting with the requested month/week and year
}
\description{
\code{subset.sql.mydata} selects \code{ntps} time points from the entire data set starting
with the year and month or week selected by the user
}
\examples{
subset.mydata(year.start = 2010, cadence.start = 1, ntps = 12, cases = mydata)


}
