% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fxns.R
\name{SummaryByCountry}
\alias{SummaryByCountry}
\title{Disease Summary By Country}
\usage{
SummaryByCountry(diseases = "all", countries = "all",
  DICE_db = "predsci")
}
\arguments{
\item{diseases}{Can be a vector of diseases c("dengue", "flu") or "all"}

\item{countries}{Can be a vector of countries (or ISO-2 abbreviations) c("Brazil", "MX", "Thailand") or "all".  Not case sensitive.}
}
\value{
A list of diseases.  Each disease entry contains a sub-list of countries.  Each country contains total regions, regions-by-level, and min/max data-dates by cadence/level.
}
\description{
Generate a summary of each disease/country/level available in the database.  Attributes returned include number of regions per level, data cadence, and min/max dates by country/level/cadence.
}
