% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_fxns.R
\name{synthetic.compare}
\alias{synthetic.compare}
\title{Plot Posterior Distributions for a Synthetic Example
For each region we plot in blue the posterior density of R0 and of pC.  In red we show the input value
The entire history of the chain is used when showing the density.
In the case of a coupled model we also compare the values for the saturation distance and the
distance power (s_d and gamma respectively).}
\usage{
synthetic.compare(tab.fit = NULL, tab.cpl = NULL, fit_par = NULL,
  cpl_par = NULL, mydata = mydata, opt.list = NULL,
  run.list = NULL, imask = NULL, ireal = 1)
}
\arguments{
\item{tab.fit}{The MCMC history of the fit}

\item{tab.cpl}{MCMC history of the coupling parameters (if present)}

\item{fit_par}{Parameter values used to create the synthetic profiles}

\item{cpl_par}{Coupling parameter values used to create the synthetic profile}

\item{mydata}{a list with all the data available for this run}

\item{opt.list}{A logical list of all the parameters \pkg{DICE} recognizes and
can optimize with TRUE/FALSE}

\item{run.list}{a list with parameters used for the MCMC procedure}

\item{imask}{An array of integers with +1/-1 values for parameters that are optimized (or not)}

\item{ireal}{Integer - the MCMC chain number}
}
\value{
err   Returns \eqn{err = 0}
}
\description{
Plot Posterior Distributions for a Synthetic Example
For each region we plot in blue the posterior density of R0 and of pC.  In red we show the input value
The entire history of the chain is used when showing the density.
In the case of a coupled model we also compare the values for the saturation distance and the
distance power (s_d and gamma respectively).
}
\examples{
synthetic.compare{tab.fit = tab.fit, tab.cpl = tab.cpl, fit_par = save_fit_par, cpl_par = save_cpl_par,opt.list = opt.list,
run.list = run.list, mydata = mydata, imask = imask, ireal = 1}
}
