% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corTable}
\alias{corTable}
\title{Pearson Correlation Table for \% ILI}
\format{An object of class \code{matrix} with 132 rows and 6 columns.}
\usage{
corTable
}
\description{
A numeric correlation table with 132 rows and 6 columns showing for each season and each HHS region the season that has the highest correlation with it.
This season is referred to as the 'prior season'
Correlation is also shown for the national data- which is denoted as region 11.
\describe{
  \item{Region}{Starting year for the prior flu season for regions 1-10 and national - 11}
  \item{year.start}{Starting year for the flu season}
  \item{year.end}{Ending year for the flu season}
  \item{year.start.prior}{Starting year for the prior flu season}
  \item{year.end.prior}{Ending year for the prior flu season}
  \item{Cor}{The Pearson correlation between the two seasons}
  ...
}
}
\keyword{datasets}
