% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll}
\alias{roll}
\title{a function for rolling dice!}
\usage{
roll(times = 1, sides = 6, verbose = TRUE)
}
\arguments{
\item{times}{integer, number of times to roll}

\item{sides}{integer, number of sides}

\item{verbose}{logical, whether to show what was rolled}
}
\value{
a random number
}
\description{
a function for rolling dice!
}
\examples{
roll()
}
