% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll}
\alias{roll}
\alias{roll2}
\title{Roll some dice}
\usage{
roll(times = 1, sides = 6, verbose = TRUE)

roll2()
}
\arguments{
\item{times}{integer}

\item{sides}{integer}

\item{verbose}{logical}
}
\value{
A vector of results
}
\description{
A function for rolling dice
}
\details{
roll is a generic function, roll2 rolls 2d6
}
\examples{
roll()
}
